requirements.txt
# Optional dependency for enhanced template parsing
# Install with: pip install mwparserfromhell
mwparserfromhell>=0.6.4

# Note: The extractor works without external dependencies
# using only Python standard library, but mwparserfromhell
# provides better template parsing if available.

# YAML support for reading/writing dictionaries
pyyaml>=6.0.1